feat: amendment modules create-delete form page

+++ src/client/actions/amendment-actions.ts
@@ -1,9 +1,23 @@
 import { env } from "../helpers/env";
-import { CreateAmendment, CreateAmendmentParams } from "../schema/amendment";
+import {
+  CreateAmendment,
+  AmendmentByIdParams,
+  AmendmentByIdParamsSchema,
+  fetchAmendmentByIdReponse,
+  fetchAmendmentByIdReponseSchema,
+} from "../schema/amendment";
 
 export class AmendmentActions {
-  static async REMOVE(params: CreateAmendmentParams) {
-    const URL = `${env.API_BASE_URL}/api/contracts/${params.contractId}/amendments/${params.amendmentId}`;
+  static async REMOVE(params: AmendmentByIdParams) {
+    const parsedParams = AmendmentByIdParamsSchema.safeParse(params);
+
+    if (!parsedParams.success) {
+      throw new Error(parsedParams.error.message);
+    }
+
+    const { amendmentId, contractId } = parsedParams.data;
+
+    const URL = `${env.API_BASE_URL}/api/contracts/${contractId}/amendments/${amendmentId}`;
 
     const response = await fetch(URL, {
       method: "DELETE",
@@ -35,5 +49,40 @@ export class AmendmentActions {
 
     const result = await response.json();
     return result;
+  }
+
+  static async FETCH(params: AmendmentByIdParams) {
+    const parsedParams = AmendmentByIdParamsSchema.safeParse(params);
+
+    if (!parsedParams.success) {
+      throw new Error(parsedParams.error.message);
+    }
+
+    const { amendmentId, contractId } = parsedParams.data;
+
+    const response = await fetch(
+      `${env.API_BASE_URL}/api/contracts/${contractId}/amendments/${amendmentId}`,
+      {
+        cache: "no-store",
+      }
+    );
+
+    if (!response.ok) {
+      throw new Error(response.statusText);
+    }
+
+    const json: fetchAmendmentByIdReponse = await response.json();
+
+
+
+    const parsedReponse = fetchAmendmentByIdReponseSchema.safeParse(json);
+
+    if (!parsedReponse.success) {
+      throw new Error(parsedReponse.error.message);
+    }
+
+        console.log(parsedReponse.data);
+
+    return parsedReponse.data;
   }
 }
